+++ apps/sb-vite/tamagui.config.ts
@@ -1,0 +1,8 @@
+import { config } from '@tamagui-extras/config'
+
+export type Conf = typeof config
+
+declare module 'tamagui' {
+  interface TamaguiCustomConfig extends Conf {}
+}
+export default config