+++ apps/next/CHANGELOG.md
@@ -1,5 +1,13 @@
 # next-app
 
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/config@1.6.0
+  - app@0.0.25
+
 ## 0.1.11
 
 ### Patch Changes