+++ packages/app/CHANGELOG.md
@@ -1,5 +1,20 @@
 # app
 
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @tamagui-extras/rich-text@1.7.0
+  - @tamagui-extras/youtube@1.7.0
+  - @tamagui-extras/config@1.6.0
+  - @tamagui-extras/video@1.8.0
+  - @tamagui-extras/core@1.7.0
+  - @tamagui-extras/date@1.9.0
+  - @tamagui-extras/file@1.7.0
+  - @tamagui-extras/form@1.10.0
+  - @tamagui-extras/link@1.7.0
+
 ## 0.0.24
 
 ### Patch Changes