+++ packages/link/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tamagui-extras/link
 
+## 1.7.0
+
+### Minor Changes
+
+- update tamagui deps and fix a month breakpoint issue when custom breakpoints are set in the Tamagui theme
+
 ## 1.6.0
 
 ### Minor Changes